+++ packages/common/core/src/toggle/toggle.machine.ts
@@ -1,0 +1,60 @@
+// Modified from https://github.com/chakra-ui/zag
+// MIT License
+// Changes from Qualcomm Technologies, Inc. are provided under the following license:
+// Copyright (c) Qualcomm Technologies, Inc. and/or its subsidiaries.
+// SPDX-License-Identifier: BSD-3-Clause-Clear
+
+import {createMachine, type MachineConfig} from "@qualcomm-ui/utils/machine"
+
+import type {ToggleSchema} from "./toggle.types"
+
+export const toggleMachine: MachineConfig<ToggleSchema> =
+  createMachine<ToggleSchema>({
+    actions: {
+      setPressed({context, event}) {
+        if ("value" in event) {
+          context.set("pressed", event.value || false)
+        }
+      },
+      togglePressed({context}) {
+        context.set("pressed", !context.get("pressed"))
+      },
+    },
+
+    context({bindable, prop}) {
+      return {
+        pressed: bindable<boolean>(() => ({
+          defaultValue: prop("defaultPressed"),
+          onChange(value) {
+            prop("onPressedChange")?.(value)
+          },
+          value: prop("pressed"),
+        })),
+      }
+    },
+
+    initialState() {
+      return "idle"
+    },
+
+    on: {
+      "PRESS.SET": {
+        actions: ["setPressed"],
+      },
+      "PRESS.TOGGLE": {
+        actions: ["togglePressed"],
+      },
+    },
+
+    props({props}) {
+      return {
+        defaultPressed: false,
+        dir: "ltr",
+        ...props,
+      }
+    },
+
+    states: {
+      idle: {},
+    },
+  })