+++ packages/frameworks/react-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @qualcomm-ui/react-core
 
+## 1.1.1 (2025/12/24)
+
+### Miscellaneous Chores
+
+- **deps:** update dependencies [@qualcomm-ui/core]
+
 ## 1.1.0 (2025/12/11)
 
 ### Features